refactor SQLRewriteEntry
